moved things into an include directory, restructured files
